// src/findjobnu-api/models/Education.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FindjobnuService
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Profile } from './Profile';
import {
    ProfileFromJSON,
    ProfileFromJSONTyped,
    ProfileToJSON,
    ProfileToJSONTyped,
} from './Profile';

/**
 * 
 * @export
 * @interface Education
 */
export interface Education {
    /**
     * 
     * @type {number}
     * @memberof Education
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Education
     */
    institution?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Education
     */
    degree?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Education
     */
    fromDate?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Education
     */
    toDate?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Education
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Education
     */
    linkedinUrl?: string | null;
    /**
     * 
     * @type {number}
     * @memberof Education
     */
    profileId?: number;
    /**
     * 
     * @type {Profile}
     * @memberof Education
     */
    profile?: Profile;
}

/**
 * Check if a given object implements the Education interface.
 */
export function instanceOfEducation(value: object): value is Education {
    return true;
}

export function EducationFromJSON(json: any): Education {
    return EducationFromJSONTyped(json, false);
}

export function EducationFromJSONTyped(json: any, ignoreDiscriminator: boolean): Education {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'institution': json['institution'] == null ? undefined : json['institution'],
        'degree': json['degree'] == null ? undefined : json['degree'],
        'fromDate': json['fromDate'] == null ? undefined : json['fromDate'],
        'toDate': json['toDate'] == null ? undefined : json['toDate'],
        'description': json['description'] == null ? undefined : json['description'],
        'linkedinUrl': json['linkedinUrl'] == null ? undefined : json['linkedinUrl'],
        'profileId': json['profileId'] == null ? undefined : json['profileId'],
        'profile': json['profile'] == null ? undefined : ProfileFromJSON(json['profile']),
    };
}

export function EducationToJSON(json: any): Education {
    return EducationToJSONTyped(json, false);
}

export function EducationToJSONTyped(value?: Education | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'institution': value['institution'],
        'degree': value['degree'],
        'fromDate': value['fromDate'],
        'toDate': value['toDate'],
        'description': value['description'],
        'linkedinUrl': value['linkedinUrl'],
        'profileId': value['profileId'],
        'profile': ProfileToJSON(value['profile']),
    };
}